+++ CHANGELOG.md
@@ -56,7 +56,7 @@
 * Annotated methods in WebCodecs that throw.
   [#3970](https://github.com/rustwasm/wasm-bindgen/pull/3970)
 
-* Update and stabilize widely supported parts of the Clipboard API.
+* Update and stabilize the Clipboard API.
   [#3992](https://github.com/rustwasm/wasm-bindgen/pull/3992)
 
 ### Fixed